accept policies page
