+++ ui/pages/Login/reducers.js
@@ -3,15 +3,17 @@ import queryString from 'query-string'
 
 import { HttpRequestHelper } from 'shared/utils/httpRequestHelper'
 
+const redirectNext = () => {
+  // Redirect to next page or home page
+  window.location.href = `${window.location.origin}${queryString.parse(window.location.search).next || ''}`
+}
+
 // Data actions
 
 export const login = (values) => {
   return () => {
     return new HttpRequestHelper('/api/login',
-      () => {
-        // Redirect to next page or home page
-        window.location.href = `${window.location.origin}${queryString.parse(window.location.search).next || ''}`
-      },
+      redirectNext,
       (e) => {
         throw new SubmissionError({ _error: [e.message] })
       },
@@ -37,6 +39,17 @@ export const setPassword = (values) => {
         // Redirect to home page
         window.location.href = window.location.origin
       },
+      (e) => {
+        throw new SubmissionError({ _error: [e.message] })
+      },
+    ).post(values)
+  }
+}
+
+export const updateUserPolicies = (values) => {
+  return () => {
+    return new HttpRequestHelper('/api/users/update_policies',
+      redirectNext,
       (e) => {
         throw new SubmissionError({ _error: [e.message] })
       },
