+++ ui/redux/rootReducer.js
@@ -97,19 +97,6 @@ export const loadUserOptions = (analystsOnly) => {
 
 export const loadAnalystOptions = () => loadUserOptions(true)
 
-export const updateUserPolicies = (values) => {
-  return (dispatch) => {
-    return new HttpRequestHelper('/api/users/update_policies',
-      (responseJson) => {
-        dispatch({ type: UPDATE_USER, updates: responseJson })
-      },
-      (e) => {
-        throw new SubmissionError({ _error: [e.message] })
-      },
-    ).post(values)
-  }
-}
-
 export const loadProject = (projectGuid, requestType = REQUEST_PROJECTS, detailField = 'variantTagTypes') => {
   return (dispatch, getState) => {
     const project = getState().projectsByGuid[projectGuid]
